moved subcomponents to their own files
