+++ packages/editor/src/components/entities-saved-states/entity-type-list.js
@@ -1,0 +1,67 @@
+/**
+ * External dependencies
+ */
+import { some } from 'lodash';
+
+/**
+ * WordPress dependencies
+ */
+import { PanelBody } from '@wordpress/components';
+import { useSelect } from '@wordpress/data';
+import { page, layout, grid, blockDefault } from '@wordpress/icons';
+
+/**
+ * Internal dependencies
+ */
+import EntityRecordItem from './entity-record-item';
+
+const ENTITY_NAME_ICONS = {
+	site: layout,
+	page,
+	post: grid,
+	wp_template: grid,
+};
+
+export default function EntityTypeList( {
+	list,
+	unselectedEntities,
+	setUnselectedEntities,
+	closePanel,
+} ) {
+	const firstRecord = list[ 0 ];
+	const entity = useSelect(
+		( select ) =>
+			select( 'core' ).getEntity( firstRecord.kind, firstRecord.name ),
+		[ firstRecord.kind, firstRecord.name ]
+	);
+
+	// Set icon based on type of entity.
+	const { name } = firstRecord;
+	const icon = ENTITY_NAME_ICONS[ name ] || blockDefault;
+
+	return (
+		<PanelBody title={ entity.label } initialOpen={ true } icon={ icon }>
+			{ list.map( ( record ) => {
+				return (
+					<EntityRecordItem
+						key={ record.key || 'site' }
+						record={ record }
+						checked={
+							! some(
+								unselectedEntities,
+								( elt ) =>
+									elt.kind === record.kind &&
+									elt.name === record.name &&
+									elt.key === record.key
+							)
+						}
+						onChange={ ( value ) =>
+							setUnselectedEntities( record, value )
+						}
+						closePanel={ closePanel }
+					/>
+				);
+			} ) }
+		</PanelBody>
+	);
+}
